feat: lodash 를 dependency 에 추가
- lodash 에서 한개의 모듈을 꺼내와서 Import 하면

Module not found: Error: Can't resolve 'lodash/sortBy' in '/Users/gaki2/Documents/tft-docs/packages/tft-ui/lib/components/team_builder/champion_list/base'
Did you mean 'sortBy.js'?
BREAKING CHANGE: The request 'lodash/sortBy' failed to resolve only because it was resolved as fully specified
(probably because the origin is strict EcmaScript Module, e. g. a module with javascript mimetype, a '*.mjs' file, or a '*.js' file where the package.json contains '"type": "module"').
The extension in the request is mandatory for it to be fully specified.
Add the extension to the request.
위와 같은 에러가 떠서 일단 _ 로 통째로 임포트해서 사용했음..